wb/requirements.txt
a==1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.1
bcrypt==3.2.2
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
decorator==5.1.1
Deprecated==1.2.13
discord==1.7.3
discord.py==1.7.3
distlib==0.3.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
inflect==6.0.0
install==1.3.5
itsdangerous==2.0.1
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.0
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
more-itertools==8.14.0
multidict==6.0.2
of==1.0.1
packaging==21.3
Pillow==9.2.0
portend==3.1.0
pycparser==2.21
pydantic==1.9.2
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
SQLAlchemy==1.4.40
tempora==5.0.2
to==0.3
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
ws4py==0.5.1
WTForms==3.0.1
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.1